Added a GUI class and separated the gui elements from the main class and removed extra imports
